% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_advanced_stats.R
\name{pfr_advanced_stat_season}
\alias{pfr_advanced_stat_season}
\alias{pfr_advanced_receiving_season}
\alias{pfr_advanced_rushing_season}
\alias{pfr_advanced_defense_season}
\title{Scrape PFR Advanced Stats on Season Level}
\usage{
pfr_advanced_stat_season(s, type = c("receiving", "rushing", "defense"))

pfr_advanced_receiving_season(s)

pfr_advanced_rushing_season(s)

pfr_advanced_defense_season(s)
}
\arguments{
\item{s}{Season to scrape}

\item{type}{Stat type}
}
\value{
A tibble
}
\description{
Scrape PFR Advanced Stats on Season Level
}
