% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_advstats.R
\name{pfr_game_adv_stats}
\alias{pfr_game_adv_stats}
\title{Scrape advanced stats from game}
\usage{
pfr_game_adv_stats(game_id)
}
\arguments{
\item{game_id}{PFR game ID}
}
\value{
a list of four dataframes: passing,rushing, receiving, and defense
}
\description{
Scrape advanced stats from game
}
